requirements.txt
python-dotenv==1.0.0
psycopg[binary,pool]==3.1.10
django==4.2.1
djangorestframework==3.14.0
djoser==2.2.0
drf-yasg==1.21.7
django-filter==23.2
django-cors-headers==4.2.0
mysqlclient
